Updated "EP-Fetch2" & "Example Package" compile scripts to validate for Mpm V1.5 before compiling.

+++ z88apps/epfetch/make.epfetch.sh
@@ -16,6 +16,17 @@
 #
 # *************************************************************************************
 
+# return version of Mpm to command line environment.
+# validate that MPM is V1.5 or later - only this version or later supports macros
+MPM_VERSIONTEXT=`mpm -version`
+
+if test $? -lt 15; then
+  echo Mpm version is less than V1.5, Ep-Fetch2 compilation aborted.
+  echo Mpm displays the following:
+  mpm
+  exit 1
+fi
+
 # compile EazyLink application from scratch
 rm -f *.obj *.bin *.map
 mpm -bg -I../../oz/def epfetch2
